+++ spec/app/models/template.rb
@@ -1,0 +1,5 @@
+class Template
+  include Mongoid::Document
+  field :active, type: Boolean, default: false
+  validates :active, presence: true
+end